+++ src/collective/documentgenerator/content/pod_template.py
@@ -299,7 +299,7 @@ class IConfigurablePODTemplate(IPODTemplate):
 
     form.widget('merge_templates', DataGridFieldFactory)
     merge_templates = schema.List(
-        title=_(u'Templates to merge.'),
+        title=_(u'Templates to merge'),
         required=False,
         value_type=DictRow(
             schema=IMergeTemplatesRowSchema,