+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog
 -----------------
 
 - Added views to convert an odt file to PDF.
-  [chris-adam]
+  [chris-adam, sgeulette]
 
 
 3.45 (2025-06-25)