+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 3.46 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Added views to convert an odt file to PDF.
+  [chris-adam, sgeulette]
 
 
 3.45 (2025-06-25)